feat: improve blockchain mining algorithm
Refactor the `mine-block` function in `core.clj` to improve the blockchain mining algorithm. The code changes introduce a nested `mine` function that performs the mining process. It checks if the hash of the blockchain starts with "0000" and updates the blockchain's hash accordingly. This change enhances the security and reliability of the blockchain.